src/test/java/com/xq/sonarqubedemo/representation/controller/PeopleControllerTest.java
package com.xq.sonarqubedemo.representation.controller;

import com.xq.sonarqubedemo.application.service.PeopleApplicationService;
import com.xq.sonarqubedemo.domain.entity.People;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import java.util.Collections;

import static org.hamcrest.Matchers.hasSize;
import static org.mockito.Mockito.when;
import static org.springframework.http.MediaType.APPLICATION_JSON;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(PeopleController.class)
public class PeopleControllerTest {

    @Autowired
    private MockMvc mockMvc;
    @MockBean
    private PeopleApplicationService service;

    @Test
    public void shouldFetchAllPeoples() throws Exception {
        when(service.fetchAll()).thenReturn(Collections.singletonList(new People("cxq")));

        mockMvc.perform(get("/peoples/").accept(APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(1)));
    }
}